pyproject.toml
[tool.black]
line-length = 100
target-version = ['py39']
quiet = true
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | build
  | dist
  | [a-z_]+/migrations
)/
'''

[tool.flake8]
max-line-length = 100
exclude = [
    '*/migrations/*', 
    '*/views.py',
    '*/apps.py',
    'venv/*', 
    'manage.py',
]
per-file-ignores = [
    '*/tests/__init__.py: F401', 
    '*/settings/*: F405',
    '__init__.py: F401, F403',
]

[tool.poetry]
name = "rnr rooms"
version = "0.1.0"
description = ""
authors = ["P&E <tech@aamarpay.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
django-admin = "^2.0.2"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
hypercorn = "^0.15.0"
psycopg2 = "^2.9.9"
drf-spectacular = "^0.27.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.11.0"
django-stubs = "^4.2.6"
ipython = "^8.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
